Remove guides in client

+++ Code/ProjectMain/Client/src/main/java/org/supportmeinc/view/ThumbnailItem.java
@@ -26,10 +26,11 @@ public class ThumbnailItem {
 
     private UUID guideUUID;
     private GuideBrowser guideBrowser;
+    private boolean author;
 
-    public void setData(String title, byte[] image, String text, UUID guideUUID, GuideBrowser guideBrowser) {
+    public void setData(String title, byte[] image, String text, UUID guideUUID, GuideBrowser guideBrowser, boolean author) {
         this.guideUUID = guideUUID;
-
+        this.author = author;
         lblTitle.setText(title);
         imgThumb.setImage(ImageUtils.toImage(image));
         lblText.setText(text);
@@ -65,8 +66,10 @@ public class ThumbnailItem {
         });
 
         contextMenu.getItems().add(openItem);
-        contextMenu.getItems().add(editItem);
-        contextMenu.getItems().add(deleteItem);
+        if (author) {
+            contextMenu.getItems().add(editItem);
+            contextMenu.getItems().add(deleteItem);
+        }
         contextMenu.getItems().add(cancelItem);
 
         anchPane.setOnContextMenuRequested(new EventHandler<ContextMenuEvent>() {
@@ -97,9 +100,9 @@ public class ThumbnailItem {
         alert.setY(anchPane.getLayoutY());
         Optional<ButtonType> result = alert.showAndWait();
         if(result.isEmpty() || result.get() != ButtonType.OK) {
-            System.out.println("dont delete");
         } else {
             System.out.println("do delete");
+            guideBrowser.deleteGuide();
         }
 
     }
@@ -109,7 +112,7 @@ public class ThumbnailItem {
     }
 
     public void setSelectedGuide() {
-        guideBrowser.setCurrentUUID(guideUUID);
+        guideBrowser.setSelectedThumbnail(guideUUID, author);
         anchPane.setStyle("-fx-border-color: black");
     }
 }
